more lint warnings

+++ projects/Standard/src/core/collections/SingleLinkedListCorrectIterOrderGeneric.java
@@ -69,7 +69,7 @@ public class SingleLinkedListCorrectIterOrderGeneric<T> {
 		SingleLinkedListCorrectIterOrderGeneric.Iterator<Integer> it = li
 				.getIterator();
 		while (it.hasNext()) {
-			Integer i = (Integer) it.next();
+			Integer i = it.next();
 			System.out.println("i is " + i);
 		}
 	}
